Def of 'order' was wrong
This should basically always be 1 now, but the first (small scale) clumping algorithm was getting confused, and assigning balls higher order than they deserved. This meant that the ODE being solved was much higher order than was necessary (although this should have not affected the results, as everything scaled the same).