Def of 'order' was wrong
This should basically always be 1 now, but the first (small scale) clumping algorithm was getting confused, and assigning balls higher order than they deserved. This meant that the ODE being solved was much higher order than was necessary (although this should have not affected the results, as everything scaled the same).

Former-commit-id: 559a19f1f3f99d735bc717cb6b1248781aabe5fa [formerly 82752216b407c6f75db1a4f93fb013f470788840] [formerly 0667e4f8243bd502358e16a03692ce21b122d7e5]
Former-commit-id: ac89bf9ae8439d250cb7550e396a990fad8503cc
Former-commit-id: f42b21cd5e6990c98f5c231d91171d76d51f121e
Former-commit-id: 9240ca3e7edc5e01cba0ea6b4c70daf97d4f893a
Former-commit-id: 83d21673ffb09f390fd74d1e8bed604094acfe00