Def of 'order' was wrong
This should basically always be 1 now, but the first (small scale) clumping algorithm was getting confused, and assigning balls higher order than they deserved. This meant that the ODE being solved was much higher order than was necessary (although this should have not affected the results, as everything scaled the same).

Former-commit-id: 0667e4f8243bd502358e16a03692ce21b122d7e5
Former-commit-id: 82752216b407c6f75db1a4f93fb013f470788840